screen-ls: fix handling of screen exit status
The previous version of the 'screen-ls' program had code left in that was
being used by the author to verify error handling by forcing the exception
condition, with the result that the program was unusable. This changeset
replaces all invocations of:

    screen -q -ls

with:

    screen -ls

and "corrects" the detection of the screen exit status (in $PIPESTATUS) to
check for 1 (determined by empirical testing) rather than >= 9 (which was
correct for the 'screen -q' behavior; see screen(1) for details)). This is a
dupious correction, though, as screen(1) does not make clear what the expected
exit status of 'screen -ls' should be; it's conceivable (hell, downright
likely) that the 'screen' program will exit with status 1 for other errors, as
well, and since it's not defined in screen(1), it's also conceivable that
'screen -ls' could exit with some status other than 1. This makes it difficult
to implement robust error handling from programs such as 'screen-ls' that
invoke screen(1). There's a long standing bug in the uptream screen(1) issue
tracker to have this corrected:

    "bug #26750: unsuccessful exit status when -list successfully matches a socket"
    http://savannah.gnu.org/bugs/?26750

The screen-ls program has no choice but to treat exit status 1 from 'screen
-ls' as a non-error -- otherwise the program simply cannot work at all -- so
that's what we're doing for now...

Also updated the docs.