add file: .gitattributes
Introducing a .gitattributes file with 'linguist-*' attributes in it allows us
some control over how GitHub computes language stats for the project.

GitHub uses the 'linguist' library to compute statistics for the languages in
the files a project. To allow authors to override the default effect, it
honors the following attributes: linguist-documentation, linguist-language,
and linguist-vendored. Documentation and "vendored" files do not contribute to
the language statistics for a repo.

See also:  https://github.com/github/linguist#troubleshooting