fixes #2: extract session names from `'screen -ls'` using `grep`
Prior to this changeset we used a combination of `head(1)` and `tail(1)` to
extract from the output of `'screen -ls'` just those lines that contained
session names. The `head` invocation, in particular (which was used to always
chop off the last two lines unconditionally).

That approach worked fine in older versions of `screen(1)`, but can no longer
be assumed. Testing with:
```
    $ screen --version
    Screen version 4.00.03 (FAU) 23-Oct-06
```
and:
```
    $ screen --version
    Screen version 4.03.01 (GNU) 28-Jun-15
```
the former emits an extra newline char after the "N Sockets in /var/run/..."
message and the latter does not.

This changeset modifies the `screen-ls` program to use `grep(1)` to extract
from the `'screen -ls'` output all non-empty lines that begin with a space,
and removes the invocations of `head(1)` and `tail(1)`. This approach should
be a bit less brittle.