[1 of 2] fixes #4 (build): generate html-formatted man pages
The project is now able to generate man pages in HTML format (in addition to
the troff format man pages). This is mainly intended for use by the author to
incorporate them into the `'screen-wrappers'` project's GitHub Pages web site.
However, the mechanism is generic and will install the pages beneath a
`'html'` directory beneath the `htmldir` location (which can be specified at
configure-time).

Since this is a new capability, we have bumped up the 'm' in our M.m.p
(semver) version number from `'0.2.3-snapshot'` to `'0.3.0-snapshot'`.