README.md: urls are not code blocks
There were a number of places where we have stand-alone URLs on a line, but
because they had four leading spaces the markdown interpreters are treating
them as code blocks. This changeset removes one leading space to prevent URLs
from being interpretted as code blocks (when the markdown is rendered as
HTML).