screen-wrappers-0.3.0 release prep
Updated version in 'configure.ac', set the release date in 'NEWS', and
regenerated 'configure'.

Also added a rule to src/main/bash/bin/Makefile.am to remove pod2html
temporary cache files. This should not be necessary since 2011 or so, but is
for me; I don't have time at the moment to track it all the way back.

    https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=616582
    https://rt.perl.org/Public/Bug/Display.html?id=69782