Add structure for generated man pages
For now, the cookbook was to configure the release source tarball with:

    --htmldir=/path/to/wherever

and then copy the generated files into our gh-pages tree here. I made a few
manual edits to the files (nothing crazy) before committing them.

At the moment there are really only two files involved:

    man/man1/screen-ls.1.html
    stylesheets/pod2html-manpage.css

I've updated the top-level index.html so that "screen-ls(1)" references are
now HTML anchors that point to the above page.

I'd much prefer a fully automated solution, but it's a start.