[1 of 2] fixes #5 (build): remove '--nobacklink' from pod2html invocation
The `'--no-backlink'` option to `pod2html` requests what is the default
behavior for that tool, but that command line option is not available on some
older `pod2html` versions (such as that from perl 5.8.8) on hosts on which we
wish to build screen-wrappers. We therefore remove the option from the
Makefile recipe with this changeset; the resulting build is identical in
behavior and outcome to that prior to this change, but the build is now able
to complete successfully on a slightly wider range of machines (at the expense
of the build recipe being slightly less explicit about the wanted behavior).