fixes #7 (screen-ls): program no longer swallows output from old screen versions
Prior to this changeset the `'screen-ls'` program was only able to parse the
`'screen -ls'` output from more recent versions of GNU `'screen'`, and was
swallowing all of the output from ancient versions of `'screen'` that produced
a slightly different output.

The program was expecting to parse output in the form:
```
        28189.screen-wrappers-hacking   (06/22/2016 06:49:18 PM)        (Attached)
```
but was inadvertently swallowing all of the output from older versions, which
have the form:
```
        28189.screen-wrappers-hacking   (Attached)
```
The `screen-wrappers` author is still encountering those ancient versions "in
the wild", so the `'screen-ls'` program has been enhanced to be flexible
enough to correctly parse both forms.

Old format behavior tested with `screen` 4.00.03 (FAU) 23-Oct-06.

New format behavior tested with `screen` 4.04.00 (GNU) 19-Jun-16.