fixes #6 (build): have configure check for bash 4.x or newer
The `'screen-ls'` program uses associative arrays, which were introduced in
`bash-4.0`. The `'configure'` was previously checking for the presence of
`bash`, but did not check for any particular version. The consequence was that
it was possible to build a version of the program that did not work at runtime
if installing on a host that only has bash 3.x (yes, the author came across
such a thing in the wild).

The `'configure'` script now checks that the found version of bash is at major
version 4 or newer, and bombs out with an error message if that is not the
case.

If you have multiple versions of bash installed on your machine, you can tell
`'configure'` where to find the one you want to use by specifying the
`'BASH_SH'` variable on the `'configure'` command line like this:
```
    $ ./configure --prefix=/path/to/wherever BASH_SH=/path/to/your/bash-4
```