fixes #10: update 'BUGS' file to direct user to GitHub issues page
Prior to this changeset the text of the 'BUGS' file directed the user to send
bug reports directly to the author via email. While that approach is fine as a
fallback, the preferred method of reporting bugs is via the project's GitHub
issue tracker page. That is likely to make them stand out so they get the
attention they need from the author (me), and also helps ensure that they
won't be lost (in my email inbox) if the project is transferred to somebody
else in the future.