fixes #9 (screen-ls): program now handles "no existing screen sessions found" scenario
Prior to this changeset, the `'screen-ls'` program was not correctly detecting
the absence of any screen sessions.

This changeset adds logic to check exit status of `'screen -q -ls'` prior to
attempting to parse the output of `'screen -ls'`. This provides a more direct
way of detecting the "no screen sessions found" scenario because the `'-q'`
option causes `'screen'` to exit with status `9`. The `'screen'` exit statuses
between `10` and `125` are treated as "one or more screen session found". The
behavior of exit status `10` and above is consistent with `screen(1)`, which
states:

>      -q   Suppress printing of error messages. In combination with "-ls" the
>           exit  value  is  as  follows: 9 indicates a directory without ses-
>           sions. 10 indicates a directory with running  but  not  attachable
>           sessions.  11 (or more) indicates 1 (or more) usable sessions.  In
>           combination with "-r" the exit value is as follows:  10  indicates
>           that  there  is  no session to resume. 12 (or more) indicates that
>           there are 2 (or more) sessions to resume and  you  should  specify
>           which one to choose.  In all other cases "-q" has no effect.

[The above taken from `screen(1)` as provided with screen version 4.00.03 (FAU)
23-Oct-06. Note that the man page text for the option remains identical to the
above as of screen version 4.05.00 (GNU) 10-Dec-16]

The limit of treating `125` as the highest exit status interpreted as
described above is due to the fact that `'bash'` uses values of `126` and
higher to indicate various abnormal exit conditions (killed by various
signals, etc.). For more on that, see the "EXIT STATUS" section of `bash(1)`.

Consistent with prior releases of the `'screen-wrappers'` package, the
`'screen-ls'` program still exits with a zero (success) status if it was able
to complete its work successfully, regardless of whether any existing screen
sessions were found.

If verbose output has been enabled, then a message is printed to `stderr`
explaining the "no screen sessions found" scenario:
```sh
   $ screen-ls
   $ echo $?
   0

   $ screen-ls -v
   screen-ls (info): no screen sessions exist (okay); exiting early with success status (no work left to do)
   $ echo $?
   0
```

Note that there is now an additional invocation of the `'screen'` executable
for the purposes of obtaining the exit status of `'screen -q -ls'`. This
creates an inherent race condition in the scenario in which the first
invocation detects one or more existing screen sessions: hypothetically those
`'screen'` sessions could go away prior to the later invocation of `'screen -ls'`.

The program attempts to deal with this scenario by attempting to detect the
"no existing screen sessions found" scenario _not only_ via the exit status of
the `'screen -q -ls'` command, but _also_ when parsing the output of the
`'screen -ls'` command later. This complicates the code a little, but not too
badly; see the comments in the code for details.