screen-ls.in: remove variable $DATA_SCREEN_Q_LS_OUTPUT
The `$DATA_SCREEN_Q_LS_OUTPUT` variable was introduced recently to capture the
output of the `'screen -q -ls'` command, mainly because the code pattern was
copied from the similar setting of the `$DATA_SCREEN_LS_OUTPUT` variable when
capturing the output of the `'screen -ls'` command.

However, the `'screen -q -ls'` command does not emit any output on `stdout`,
and we have no need to confirm that (via `test -z ...` or similar), so I'm
removing the setting of `$DATA_SCREEN_Q_LS_OUTPUT` (which was not referenced
anywhere).

The program still invokes `'screen -q -ls'` and checks its exit status; it
just does not needlessly capture its (nonexistent) output.