fixes #15: screen-ls: explicitly initialize all bash arrays
In order to allow 'screen-ls' to work with bash-4.x older than bash-4.4-alpha,
we explicitly initialize all bash array variables (for both indexed arrays and
associative arrays), even when the array is empty.

Motivation: The 'screen-ls' program relies on the output of bash's 'declare -p'
builtin, and it wasn't until bash-4.4-alpha that the output would work for
variables that had been declared but not initialized (which are technically
unset).

Though the specific fix needed for issue #15 was to explicitly initialize the
two recently added global associative array variables:

    SCREEN_VERSION_ARRAY_RAW
    SCREEN_VERSION_ARRAY_NUMERIC_ONLY

(both of which had been declared but not initialized), as part of the work for
addressing this issue the screen-ls program was audited to help ensure that
there are not any other bugs of this type hiding in there; all array variables
have been checked to ensure they are explicitly initialized.